requirements.txt
pyTelegramBotAPI==4.12.0
Pillow==10.0.0
face_recognition==1.3.0
numpy==1.26.2